Rename folders and file unfiled stories
