+++ tests/06-Comments/Comments.test.tsx
@@ -1,6 +1,6 @@
 import "@testing-library/jest-dom/extend-expect";
 
-import * as stories from "stories/06-Comments/02-ChainComment.stories";
+import * as stories from "stories/Comments/02-ChainComment.stories";
 
 import { render, screen, waitFor } from "@testing-library/react";
