+++ tests/Header.test.tsx
@@ -1,6 +1,6 @@
 import "@testing-library/jest-dom/extend-expect";
 
-import * as stories from "stories/05-Posts/02-PostHeader.stories";
+import * as stories from "stories/Post/02-PostHeader.stories";
 
 import {
   RenderResult,
